CA-603 Enable dynamic bandwidth
